Solution to Print Name, Age, and Hobby

+++ Print Name, Age, and Hobby@@6788d5012db25efec946f945/code.c
@@ -2,10 +2,10 @@
 
 int main(){
     char name[50];
-    float age;
+    int age;
     char hobby[50];
-    scanf("%s%f%s",name, &age, hobby);
-    scanf("%f",&age);
-    print("\nName: %s\nAge: %f\nHobby: %s",name, age, hobby);
+    scanf("%s%d%s",name, &age, hobby);
+    scanf("%d",&age);
+    print("\nName: %s\nAge: %d\nHobby: %s",name, age, hobby);
     return 0;
 }
